fieldbioinformatics/requirements.txt
biopython
clint
pandas
pysam
pytest
pyvcf
requests
tqdm